chore: refactor, document and update build/setup process
This commit includes the following changes:

- Update the python version to 3.9 in .python-version
- Add a launch configuration for VSCode in .vscode/launch.json
- Add the GPLv3 License to the project in LICENSE.md
- Update screenshots in the README.md to have a width of 800px
- Modify build.py, pidcat.py, resources/version_info.py, setup/info.txt,
   and setup/setup.iss to ensure compatibility with Python 3.9 and higher

The changes were made to ensure that PidCat can be used with Python 3.9 and higher.
The launch configuration was added to make it easier to debug the project using VSCode.
The GPLv3 License was added to comply with the licensing requirements of the project.
Signed-off-by: AbdAlMoniem AlHifnawy <hifnawy_moniem@hotmail.com>

Change-Id: Ia819e270b861c5fb050634d223096b1803d93d87